Priviliges Escalation using Iftop with sudo
Iftop displays bandwidth usage on an interface by the host. In the previous versions of iftop, it allowed users to execute commands interactively after being invoked. If it is called with sudo, all the commands will run in context of the root user which can be used to escalate privileges.  

After getting invoked iftop (>= 1.17), displays network usage details. User can enter and execute commands interactively.