Update dosbox.md
Problems with copy and echo are solved. Echo included at the end of the data, the following Windows character: ^M, so you could not do a PrivEsc because any file you wrote, would not be "readable". 

https://imgur.com/a/creeOUU

The following can be seen with VIM
https://imgur.com/a/6WJ4b60

https://imgur.com/a/vrA9pce

More related info: https://virtuozzosupport.force.com/s/article/000014695 

Regarding copy, using type solves a bug that did not copy the content, and only inserted the following in the file: 
Copy failure : ...
   0 File(s) copied.
https://imgur.com/a/4kBM1UD