Make the SUID shell remark optional
The rationale is that if the `shell` argument is omitted, then the function just
works in any environment as SUID, hence no remark is needed. If `true` then the
example makes use of some shell invocation somewhere in the executable
implementation, that makes the function work only when the system shell does not
drop effective privileges. Finally, when set to `false` it means that the
example explicitly makes use of `/bin/sh -p`, and the executable implementation
executes the shell directly; this is most commonly found in `shell` functions.